# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fathom",
    "version": "1.0.0",
    "description": "Macroeconomic data API powered by FRED (Federal Reserve Economic Data). 816,000+ time series covering GDP, inflation, employment, interest rates, money supply, and more.",
    "x-guidance": "Use /api/popular to discover available indicators. Use /api/search to find specific series by keyword. Use /api/observations to fetch data values. Use /api/snapshot for a dashboard overview. Use /api/compare to compare multiple series. Series IDs are uppercase codes like GDP, UNRATE, CPIAUCSL, FEDFUNDS, DGS10.",
    "contact": {
      "email": "admin@orbonomy.xyz"
    },
    "x-logo": {
      "url": "https://fathom.vercel.app/favicon.png"
    }
  },
  "servers": [
    {
      "url": "https://fathom.vercel.app"
    }
  ],
  "paths": {
...
    "/api/vintage": {
      "post": {
        "operationId": "vintage",
        "summary": "Vintage dates",
        "description": "Get vintage/revision dates for a FRED series. Shows when data was revised.",
        "tags": [
          "Metadata"
        ],
        "security": [],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.020000"
          },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "network": "eip155:8453"
              }
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "seriesId": {
                    "type": "string",
                    "minLength": 1,
                    "description": "FRED series ID"
                  },
                  "limit": {
                    "type": "string",
                    "description": "Max results (default: 10)"
                  }
                },
                "required": [
                  "seriesId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vintage dates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string"
                    },
                    "series_id": {
                      "type": "string"
                    },
                    "vintage_dates": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  },
                  "required": [
                    "source"
                  ]
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
  }
}